# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Native Call Center API",
    "version": "1.0.0",
    "summary": "REST and SSE surface of the AI-native call center.",
...
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Same-origin API mount; the SPA is served by the same process."
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login",
        "tags": [
          "auth"
        ],
        "summary": "Sign in and receive the session cookie",
        "description": "Verifies the credentials and sets the HttpOnly session cookie. The response also carries the identity so the SPA can render without a second round trip.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in; Set-Cookie carries the session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "tags": [
          "auth"
        ],
        "summary": "Sign out",
        "description": "Revokes the session and clears the cookie. Always succeeds for an authenticated caller.",
        "security": [
          {
            "cookieSession": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "204": {
            "description": "Signed out."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "getMe",
        "tags": [
          "auth"
        ],
        "summary": "The signed-in identity",
        "responses": {
          "200": {
            "description": "The current identity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "apiKeyBearer": []
          }
        ]
      }
    },
    "/agent/presence": {
      "get": {
        "operationId": "getAgentPresence",
        "tags": [
          "agents"
        ],
        "summary": "The caller's own presence",
        "description": "Requires the AGENT role and an agent profile; a supervisor or administrator without one receives 403.",
        "responses": {
          "200": {
            "description": "Current presence.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Presence"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": [
          {
            "cookieSession": [
              "agent:read"
            ]
          },
          {
            "apiKeyBearer": [
              "agent:read"
            ]
          }
        ]
      }
    },
    "/agent/login": {
      "post": {
        "operationId": "agentLogin",
        "tags": [
          "agents"
        ],
        "summary": "Sign in to a phone",
        "description": "Binds the agent to an extension and enters NOT_READY (reason LOGIN). Requires the AGENT role.",
        "security": [
          {
            "cookieSession": [
              "agent:act"
            ],
            "csrfHeader": []
          },
          {
            "apiKeyBearer": [
              "agent:act"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The presence after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Presence"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/agent/logout": {
      "post": {
        "operationId": "agentLogout",
        "tags": [
          "agents"
        ],
        "summary": "Sign out of the phone",
        "description": "Requires the AGENT role.",
        "security": [
          {
            "cookieSession": [
              "agent:act"
            ],
            "csrfHeader": []
          },
          {
            "apiKeyBearer": [
              "agent:act"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "The presence after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Presence"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/agent/wrap-up": {
      "get": {
        "operationId": "getAgentWrapUp",
        "tags": [
          "agents"
        ],
        "summary": "The after-call work waiting on the caller",
        "description": "The record opened for the call this agent just finished, whether or not they have confirmed it. 204 when there is nothing to confirm. Requires the AGENT role.",
        "responses": {
          "200": {
            "description": "The open record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWrapUp"
                }
              }
            }
          },
          "204": {
            "description": "Nothing to confirm."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "cookieSession": [
              "agent:read"
            ]
          },
          {
            "apiKeyBearer": [
              "agent:read"
            ]
          }
        ]
      },
      "post": {
        "operationId": "agentWrapUp",
        "tags": [
          "agents"
        ],
        "summary": "Confirm after-call work",
        "description": "Confirms the record the platform opened when the call ended, applying whatever the agent changed, and returns them to READY. Both fields are optional — the disposition already has a value, and confirming the defaults is a legitimate answer. The filing is still accepted once the agent has moved on: the last wrapped call stays addressable until the next one starts, and 409 AGENT_NOT_IN_WRAP_UP says when there is none. Requires the AGENT role.",
        "security": [
          {
            "cookieSession": [
              "agent:act"
            ],
            "csrfHeader": []
          },
          {
            "apiKeyBearer": [
              "agent:act"
            ]
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrapUpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The presence after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Presence"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/agents": {
      "get": {
        "operationId": "listAgents",
        "tags": [
          "agents"
        ],
        "summary": "The live roster",
        "description": "Every agent with their presence, device registration and on-call flag.",
        "responses": {
          "200": {
            "description": "The roster.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "cookieSession": [
              "agent:manage"
            ]
          },
          {
            "apiKeyBearer": [
              "agent:manage"
            ]
          }
        ]
      },
      "post": {
        "operationId": "createAgent",
        "tags": [
          "agents"
        ],
        "summary": "Give a user an agent identity",
        "description": "Creates the agent identity behind an account and binds their phone.",
        "security": [
          {
            "cookieSession": [
              "config:write"
            ],
            "csrfHeader": []
          },
          {
            "apiKeyBearer": [
              "config:write"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The stored agent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/agents/{agentId}": {
      "put": {
        "operationId": "updateAgent",
        "tags": [
          "agents"
        ],
        "summary": "Rewrite an agent's configuration",
        "description": "Rebinds the phone, renames the switch identifier, or changes wrap-up.",
        "security": [
          {
            "cookieSession": [
              "config:write"
            ],
            "csrfHeader": []
          },
          {
            "apiKeyBearer": [
              "config:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The stored agent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "delete": {
        "operationId": "deleteAgent",
        "tags": [
          "agents"
        ],
        "summary": "Remove an agent identity",
        "description": "The account remains; it simply stops being an agent, and any presence is dropped.",
        "security": [
          {
            "cookieSession": [
              "config:write"
            ],
            "csrfHeader": []
          },
          {
            "apiKeyBearer": [
              "config:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removed."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aicc_session",
        "description": "Session cookie issued by POST /auth/login, for a person signed in at a browser. The cookie name is configurable (AICC_SESSION_COOKIE); aicc_session is the default. HttpOnly, SameSite=Lax. The scopes it carries are derived from the account's role at sign-in; the role itself never appears on the wire."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AICC-Csrf",
        "description": "Anti-CSRF header, required on every mutating request that authenticates with the session cookie. Any non-empty value passes; requiring a custom header forces a CORS preflight on a cross-origin caller. Missing header: 403 FORBIDDEN. It never accompanies apiKeyBearer — this header protects a credential the browser sends automatically, and a key is sent deliberately."
      },
      "apiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key, presented as `Authorization: Bearer <key>`, for a system with no browser. The key is the subject; it is not a person and never becomes one. To act for an agent — a CRM dialling for somebody on the floor — send `X-AICC-Agent-ID: <agentId>` alongside it. Without that header the subject acts for nobody, and an agent-scoped operation answers AGENT_REQUIRED. The header is refused outright on a session request (AGENT_IMPERSONATION_NOT_ALLOWED), whatever the account's role. A key carries exactly the scopes it was issued with; missing one is INSUFFICIENT_SCOPE."
      }
    },
...
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "CurrentUser": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/Identity"
          }
        },
        "required": [
          "user"
        ]
      },
...
    }
  }
}